requirements.txt
absl-py==0.13.0
alembic==1.4.3
altair==5.0.1
anyio==3.7.1
argon2-cffi==20.1.0
asgiref==3.5.0
astor==0.8.1
astroid==2.7.3
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==23.2.0
autopep8==1.5.7
Babel==2.9.0
backcall==0.2.0
backports.entry-points-selectable==1.1.1
backports.zoneinfo==0.2.1
bce-python-sdk==0.8.64
bleach==3.3.0
blinker==1.6.3
blis==0.7.4
Boruta==0.3
Brotli==1.0.9
build==1.1.1
cached-property==1.5.2
cachetools==5.3.3
catalogue==1.0.0
certifi==2024.6.2
cffi==1.14.5
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.0.0
coloredlogs==15.0.1
colorlover==0.3.0
ConfigArgParse==1.5.3
contourpy==1.0.6
cssselect==1.1.0
cssutils==2.3.0
cufflinks==0.17.3
cycler==0.11.0
cymem==2.0.5
Cython==0.29.26
databricks-cli==0.16.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
deprecation==2.1.0
dill==0.3.3
diskcache==5.2.1
distlib==0.3.4
dnspython==2.1.0
docker==5.0.3
email-validator==1.1.3
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.1.2
fastapi==0.75.2
fasttext==0.9.1
filelock==3.4.0
findspark==1.4.2
flake8==3.9.2
Flask==2.0.3
Flask-Babel==2.0.0
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
flatbuffers==24.12.23
focal-loss==0.0.7
fonttools==4.28.2
fsspec==2021.10.0
funcy==1.16
future==0.18.2
fuzzysearch==0.7.3
gadDummy @ file:///home/dsuser/rajiv/datascience/dist/gadDummy-0.1.tar.gz
gast==0.4.0
gensim==3.8.3
gevent==21.12.0
geventhttpclient==1.5.3
gibberish-detector==0.1.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.8.4
greenlet==1.1.2
grpcio==1.34.1
gunicorn==20.1.0
h11==0.14.0
h5py==3.2.1
hiredis==2.3.2
htmlmin==0.1.12
httpcore==0.17.3
httpx==0.24.1
humanfriendly==10.0
ibm-db==3.0.4
identify==2.4.0
idna==3.7
ImageHash==4.2.1
imageio==2.13.0
imbalanced-learn==0.7.0
imgaug==0.4.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel==5.5.3
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.9.3
itsdangerous==2.1.0
jedi==0.18.0
Jinja2==3.0.3
joblib==1.0.1
json5==0.9.5
jsonschema==4.17.3
jupyter-client==6.1.12
jupyter-core==4.9.1
jupyter-packaging==0.9.1
jupyter-server==1.6.2
jupyterlab==3.0.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.4.0
jupyterlab-widgets==1.0.2
keras==2.6.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
kmodes==0.11.1
Lasagne==0.1
lazy-object-proxy==1.6.0
lesscpy==0.14.0
lightgbm==3.3.0
llvmlite==0.36.0
lmdb==1.0.0
locust==2.7.3
lxml==4.6.4
Mako==1.1.6
Markdown==3.3.4
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.5.0
matplotlib-inline==0.1.2
mccabe==0.6.1
mdurl==0.1.2
minio==7.2.7
missingno==0.5.0
mistune==0.8.4
mlflow==1.20.2
mlxtend==0.19.0
mpmath==1.3.0
msgpack==1.0.3
multimethod==1.6
murmurhash==1.0.5
mxnet==1.7.0.post2
nbclassic==0.2.7
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nbqa==1.1.0
nest-asyncio==1.5.1
networkx==2.6.3
nltk==3.6.5
nodeenv==1.6.0
notebook==6.3.0
numba==0.53.1
numexpr==2.7.3
numpy==1.21.6
oauthlib==3.1.1
onnx==1.14.0
onnxruntime==1.14.1
opencv-contrib-python==4.4.0.46
opencv-python==4.11.0.86
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==21.3
paddleocr==2.3.0.2
paddlepaddle==2.1.3
pandas==1.3.5
pandas-profiling==3.1.0
pandocfilters==1.4.3
parso==0.8.2
patsy==0.5.2
petastorm==0.11.1
pexpect==4.8.0
phik==0.12.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
plac==1.1.3
platformdirs==2.3.0
plotly==5.3.1
ply==3.11
polygon==1.1.0
Polygon3==3.0.9.1
pre-commit==2.16.0
premailer==3.10.0
preshed==3.0.5
prometheus-client==0.10.1
prometheus-flask-exporter==0.18.3
prompt-toolkit==3.0.20
protobuf==3.20.2
psutil==5.9.0
ptyprocess==0.7.0
py-cpuinfo==9.0.0
py4j==0.10.9.2
pyarrow==12.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.8.1
pycaret==2.3.4
pyclipper==1.3.0
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.12.0
pydantic==1.8.2
pydeck==0.8.1b1
pyflakes==2.3.1
Pygments==2.17.2
pyLDAvis==3.2.2
pylint==2.10.2
Pympler==1.0.1
pynndescent==0.5.5
pyod==0.9.4
pyparsing==3.0.9
pyproject_hooks==1.2.0
pyrsistent==0.19.3
pyspark==3.1.2
python-dateutil==2.9.0.post0
python-editor==1.0.4
python-Levenshtein==0.12.2
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
PyYAML==5.4.1
pyzmq==22.3.0
querystring-parser==1.2.4
rapidocr-onnxruntime==1.4.4
redis==3.5.3
redis-py-cluster==2.1.3
regex==2021.10.8
requests==2.31.0
requests-oauthlib==1.3.0
rich==13.7.1
roundrobin==0.0.2
rsa==4.7.2
scikit-image==0.18.3
scikit-learn==0.24.1
scikit-plot==0.3.7
scipy==1.7.1
seaborn==0.11.1
Send2Trash==1.5.0
setuptools-scm==6.3.2
Shapely==1.8.0
shellcheck-py==0.8.0.1
six==1.16.0
smart-open==5.2.1
smmap==5.0.1
sniffio==1.3.0
spacy==2.3.7
SQLAlchemy==1.4.28
sqlparse==0.4.2
srsly==1.0.5
starlette==0.17.1
statsmodels==0.13.0
streamlit==1.23.1
sympy==1.10.1
tabulate==0.8.9
tangled-up-in-unicode==0.1.0
tenacity==8.2.3
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-addons==0.14.0
tensorflow-estimator==2.6.0
tensorflow-hub==0.12.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
textblob==0.15.3
Theano==0.8.2
thinc==7.4.5
threadpoolctl==2.1.0
tifffile==2021.11.2
tokenize-rt==4.1.0
toml==0.10.2
tomli==1.2.2
tomlkit==0.7.0
toolz==0.12.1
torch==1.8.1
torchvision==0.9.1
tornado==6.2
tqdm==4.67.1
traitlets==5.1.0
typed-ast==1.4.3
typeguard==2.12.1
typing_extensions==4.7.1
tzdata==2024.1
tzlocal==4.3.1
ultralytics==8.0.145
umap-learn==0.5.1
urllib3==1.26.7
uvicorn==0.15.0
validators==0.20.0
virtualenv==20.10.0
visions==0.7.4
visualdl==2.2.2
wasabi==0.8.2
watchdog==3.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.0.3
widgetsnbextension==3.5.1
wordcloud==1.8.1
wrapt==1.14.1
XlsxWriter==3.0.1
yapf==0.31.0
yellowbrick==1.3.post1
zipp==3.15.0
zope.event==4.5.0
zope.interface==5.4.0
